Task4 - EndpointFilter is ready!
